% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{choose}
\alias{choose}
\title{Iteration of [choose()]}
\usage{
choose(keybind, option1, option2)
}
\arguments{
\item{keybind}{Conflictive keybind}

\item{option1}{First option}

\item{option2}{Second option}
}
\value{
The chosen option
}
\description{
Iteration of [choose()]
}
\examples{
\dontrun{
 choose("Ctrl+Shift+M", "Insert magrittr pipe", "Insert pipe operator")
}
}
