% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{full_choose}
\alias{full_choose}
\title{Interactice Conflict Resolution}
\usage{
full_choose(df)
}
\arguments{
\item{df}{A merged data frame with 2 rows, with the contents of the local and cloud files}
}
\value{
df, ready to be jsoned, without conflicts
}
\description{
Used to resolve conflicts in [sync()]. Will ask the user to choose between cloud
or local, or to choose all cloud or all local.
}
\examples{
\dontrun{
  full_choose(data.frame(a = c(1,2), b = c(3,4)))
}
}
\seealso{
[sync()], [choose()]
}
