% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{progress_bar}
\alias{progress_bar}
\title{Progress Bar Closure}
\usage{
progress_bar(n)
}
\arguments{
\item{n}{Total number of iterations}
}
\value{
A function/closure that when called will update the progress bar
}
\description{
Each time the closure is called will update the progress bar
}
\examples{
 pb <- settingsSync:::progress_bar(10)
 for (i in 1:10) {
  pb()
  Sys.sleep(0.1)
 }
 pb(finish = TRUE)
}
