% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{read_from_local}
\alias{read_from_local}
\title{Read from local a JSON file inside rstudio folder}
\usage{
read_from_local(what)
}
\arguments{
\item{what}{One of c("rstudio_bindings", "editor_bindings", "addins"),
the file to read}
}
\value{
A data frame with the contents of the file, converted from json
}
\description{
This is a helper function for [sync()] that:
* reads a json file from the local rstudio folder
* converts it to a data frame (to resolve conflicts in the future)
}
\examples{
mimic_on()
  read_from_local("rstudio_bindings")
  read_from_local("editor_bindings")
  read_from_local("addins")
mimic_off()
}
\seealso{
[read_from_gd()], [push()], [sync()], [jsonlite::read_json()]
}
