% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rstudio_path}
\alias{rstudio_path}
\title{Get rstudio local path}
\usage{
rstudio_path()
}
\value{
path to rstudio local
}
\description{
Get rstudio local path
}
\examples{
rstudio_path()

}
\seealso{
[rappdirs::user_config_dir()], [is_windows()], [usethis::is_windows()]
}
