% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{pull}
\alias{pull}
\title{Pull Rstudio Settings}
\usage{
pull(addins_gd = NULL, editor_bindings_gd = NULL, rstudio_bindings_gd = NULL)
}
\arguments{
\item{addins_gd, editor_bindings_gd, rstudio_bindings_gd}{character, the json
string to be written to the respective file; if NULL, will read from
Google Drive. Can also be FALSE to not write that specific file.}
}
\description{
Pulls Rstudio settings from Google Drive, without pushing. This is just a
helper function for [sync()], the main function, but can be used alone, although
this fucntion will not do any checking, and will just override.
}
\examples{
mimic_on()
  pull()         # will pull all settings from gd, overwriting them
  pull(
   addins_gd = '{"insertPipeOperator": "Shift+Tab"}',
   editor_bindings_gd= FALSE, rstudio_bindings_gd= FALSE
  )
  # will write to addins.json the string
  pull(addins_gd = '{"insertPipeOperator": "Shift+Tab"}')
  # will write to addins.json the stringand pull the other 2 files from gd
mimic_off()
}
\seealso{
[read_from_gd()], [sync()], [push()]
}
