% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync}
\alias{sync}
\title{Sync Rstudio Settings}
\usage{
sync(write = TRUE, useProgBar = TRUE)
}
\arguments{
\item{write}{boolean, if TRUE will write the merged settings to the local files,
and push them to Google Drive. FALSE essentially just makes conflict resolution,
without changing any files (basically a dry run). Default is TRUE.}

\item{useProgBar}{boolean, if TRUE will show a progress bar. Default is TRUE.}
}
\value{
nothing
}
\description{
Gets the settings from Google Drive and from the local files, and merges them.
If there are conflicts, will ask the user to resolve them. Finally, will
write the merged settings to the local files, and push them to Google Drive.
}
\examples{
mimic_on()
if(interactive()) {
  sync(write= FALSE) # dry run, will not write to files or push to gd
  sync()             # will sync all settings, is what's run when called by addin
}
mimic_off()
}
\seealso{
[push()], [pull()]
}
